hafta_4/AkademikMakaleAsistani/requirements.txt
streamlit
chromadb
sentence-transformers
openai
PyPDF2
python-dotenv
